python/requirements.txt
grpcio
redis
